Enum polyfuse::Operation

pub enum Operation<'a> {
    Lookup {
        parent: u64,
        name: &'a OsStr,
        reply: ReplyEntry,
    },
    Forget {
        forgets: &'a [Forget],
    },
    Getattr {
        ino: u64,
        fh: Option<u64>,
        reply: ReplyAttr,
    },
    Setattr {
        ino: u64,
        fh: Option<u64>,
        mode: Option<u32>,
        uid: Option<u32>,
        gid: Option<u32>,
        size: Option<u64>,
        atime: Option<(u64, u32, bool)>,
        mtime: Option<(u64, u32, bool)>,
        ctime: Option<(u64, u32)>,
        lock_owner: Option<u64>,
        reply: ReplyAttr,
    },
    Readlink {
        ino: u64,
        reply: ReplyReadlink,
    },
    Symlink {
        parent: u64,
        name: &'a OsStr,
        link: &'a OsStr,
        reply: ReplyEntry,
    },
    Mknod {
        parent: u64,
        name: &'a OsStr,
        mode: u32,
        rdev: u32,
        umask: Option<u32>,
        reply: ReplyEntry,
    },
    Mkdir {
        parent: u64,
        name: &'a OsStr,
        mode: u32,
        umask: Option<u32>,
        reply: ReplyEntry,
    },
    Unlink {
        parent: u64,
        name: &'a OsStr,
        reply: ReplyEmpty,
    },
    Rmdir {
        parent: u64,
        name: &'a OsStr,
        reply: ReplyEmpty,
    },
    Rename {
        parent: u64,
        name: &'a OsStr,
        newparent: u64,
        newname: &'a OsStr,
        flags: u32,
        reply: ReplyEmpty,
    },
    Link {
        ino: u64,
        newparent: u64,
        newname: &'a OsStr,
        reply: ReplyEntry,
    },
    Open {
        ino: u64,
        flags: u32,
        reply: ReplyOpen,
    },
    Read {
        ino: u64,
        fh: u64,
        offset: u64,
        flags: u32,
        lock_owner: Option<u64>,
        reply: ReplyData,
    },
    Write {
        ino: u64,
        fh: u64,
        offset: u64,
        data: Bytes,
        flags: u32,
        lock_owner: Option<u64>,
        reply: ReplyWrite,
    },
    Release {
        ino: u64,
        fh: u64,
        flags: u32,
        lock_owner: Option<u64>,
        flush: bool,
        flock_release: bool,
        reply: ReplyEmpty,
    },
    Statfs {
        ino: u64,
        reply: ReplyStatfs,
    },
    Fsync {
        ino: u64,
        fh: u64,
        datasync: bool,
        reply: ReplyEmpty,
    },
    Setxattr {
        ino: u64,
        name: &'a OsStr,
        value: &'a [u8],
        flags: u32,
        reply: ReplyEmpty,
    },
    Getxattr {
        ino: u64,
        name: &'a OsStr,
        size: u32,
        reply: ReplyXattr,
    },
    Listxattr {
        ino: u64,
        size: u32,
        reply: ReplyXattr,
    },
    Removexattr {
        ino: u64,
        name: &'a OsStr,
        reply: ReplyEmpty,
    },
    Flush {
        ino: u64,
        fh: u64,
        lock_owner: u64,
        reply: ReplyEmpty,
    },
    Opendir {
        ino: u64,
        flags: u32,
        reply: ReplyOpendir,
    },
    Readdir {
        ino: u64,
        fh: u64,
        offset: u64,
        plus: bool,
        reply: ReplyData,
    },
    Releasedir {
        ino: u64,
        fh: u64,
        flags: u32,
        reply: ReplyEmpty,
    },
    Fsyncdir {
        ino: u64,
        fh: u64,
        datasync: bool,
        reply: ReplyEmpty,
    },
    Getlk {
        ino: u64,
        fh: u64,
        owner: u64,
        lk: &'a FileLock,
        reply: ReplyLk,
    },
    Setlk {
        ino: u64,
        fh: u64,
        owner: u64,
        lk: &'a FileLock,
        sleep: bool,
        reply: ReplyEmpty,
    },
    Flock {
        ino: u64,
        fh: u64,
        owner: u64,
        op: u32,
        reply: ReplyEmpty,
    },
    Access {
        ino: u64,
        mask: u32,
        reply: ReplyEmpty,
    },
    Create {
        parent: u64,
        name: &'a OsStr,
        mode: u32,
        umask: Option<u32>,
        open_flags: u32,
        reply: ReplyCreate,
    },
    Bmap {
        ino: u64,
        block: u64,
        blocksize: u32,
        reply: ReplyBmap,
    },
    Fallocate {
        ino: u64,
        fh: u64,
        offset: u64,
        length: u64,
        mode: u32,
        reply: ReplyEmpty,
    },
    CopyFileRange {
        ino_in: u64,
        off_in: u64,
        fh_in: u64,
        ino_out: u64,
        off_out: u64,
        fh_out: u64,
        len: u64,
        flags: u64,
        reply: ReplyWrite,
    },
    Poll {
        ino: u64,
        fh: u64,
        events: u32,
        kh: Option<u64>,
        reply: ReplyPoll,
    },
}

The kind of FUSE requests received from the kernel.

Variants

Lookup

Look up a directory entry by name.

Fields of Lookup

parent: u64name: &'a OsStrreply: ReplyEntry

Forget

Forget about inodes removed from the kernel's internal caches.

Fields of Forget

forgets: &'a [Forget]

Getattr

Get file attributes.

Fields of Getattr

ino: u64fh: Option<u64>reply: ReplyAttr

Setattr

Set file attributes.

Fields of Setattr

ino: u64fh: Option<u64>mode: Option<u32>uid: Option<u32>gid: Option<u32>size: Option<u64>atime: Option<(u64, u32, bool)>mtime: Option<(u64, u32, bool)>ctime: Option<(u64, u32)>lock_owner: Option<u64>reply: ReplyAttr

Readlink

Read a symbolic link.

Fields of Readlink

ino: u64reply: ReplyReadlink

Symlink

Create a symbolic link

Fields of Symlink

parent: u64name: &'a OsStrlink: &'a OsStrreply: ReplyEntry

Mknod

Create a file node.

Fields of Mknod

parent: u64name: &'a OsStrmode: u32rdev: u32umask: Option<u32>reply: ReplyEntry

Mkdir

Create a directory.

Fields of Mkdir

parent: u64name: &'a OsStrmode: u32umask: Option<u32>reply: ReplyEntry

Unlink

Remove a file.

Fields of Unlink

parent: u64name: &'a OsStrreply: ReplyEmpty

Rmdir

Remove a directory.

Fields of Rmdir

parent: u64name: &'a OsStrreply: ReplyEmpty

Rename

Rename a file.

Fields of Rename

parent: u64name: &'a OsStrnewparent: u64newname: &'a OsStrflags: u32reply: ReplyEmpty

Link

Create a hard link.

Fields of Link

ino: u64newparent: u64newname: &'a OsStrreply: ReplyEntry

Open

Open a file.

Fields of Open

ino: u64flags: u32reply: ReplyOpen

Read

Read data from an opened file.

Fields of Read

ino: u64fh: u64offset: u64flags: u32lock_owner: Option<u64>reply: ReplyData

Write

Write data to an opened file.

Fields of Write

ino: u64fh: u64offset: u64data: Bytesflags: u32lock_owner: Option<u64>reply: ReplyWrite

Release

Release an opened file.

Fields of Release

ino: u64fh: u64flags: u32lock_owner: Option<u64>flush: boolflock_release: boolreply: ReplyEmpty

Statfs

Get the filesystem statistics.

Fields of Statfs

ino: u64reply: ReplyStatfs

Fsync

Synchronize the file contents of an opened file.

When the parameter datasync is true, only the file contents should be flushed and the metadata does not have to be flushed.

Fields of Fsync

ino: u64fh: u64datasync: boolreply: ReplyEmpty

Setxattr

Set an extended attribute.

Fields of Setxattr

ino: u64name: &'a OsStrvalue: &'a [u8]flags: u32reply: ReplyEmpty

Getxattr

Get an extended attribute.

The operation should send the length of attribute's value with reply.size(n) when size is equal to zero.

Fields of Getxattr

ino: u64name: &'a OsStrsize: u32reply: ReplyXattr

Listxattr

List extended attribute names.

The attribute names must be seperated by a null character (i.e. b'\0').

The operation should send the length of attribute names with reply.size(n) when size is equal to zero.

Fields of Listxattr

ino: u64size: u32reply: ReplyXattr

Removexattr

Remove an extended attribute.

Fields of Removexattr

ino: u64name: &'a OsStrreply: ReplyEmpty

Flush

Close a file descriptor.

Fields of Flush

ino: u64fh: u64lock_owner: u64reply: ReplyEmpty

Opendir

Open a directory.

Fields of Opendir

ino: u64flags: u32reply: ReplyOpendir

Readdir

Read contents from an opened directory.

Fields of Readdir

ino: u64fh: u64offset: u64plus: boolreply: ReplyData

Releasedir

Release an opened directory.

Fields of Releasedir

ino: u64fh: u64flags: u32reply: ReplyEmpty

Fsyncdir

Synchronize an opened directory contents.

When the parameter datasync is true, only the directory contents should be flushed and the metadata does not have to be flushed.

Fields of Fsyncdir

ino: u64fh: u64datasync: boolreply: ReplyEmpty

Getlk

Test for a POSIX file lock.

Fields of Getlk

ino: u64fh: u64owner: u64lk: &'a FileLockreply: ReplyLk

Setlk

Acquire, modify or release a POSIX file lock.

Fields of Setlk

ino: u64fh: u64owner: u64lk: &'a FileLocksleep: boolreply: ReplyEmpty

Flock

Acquire, modify or release a BSD file lock.

Fields of Flock

ino: u64fh: u64owner: u64op: u32reply: ReplyEmpty

Access

Check file access permissions.

Fields of Access

ino: u64mask: u32reply: ReplyEmpty

Create

Create and open a file.

Fields of Create

parent: u64name: &'a OsStrmode: u32umask: Option<u32>open_flags: u32reply: ReplyCreate

Bmap

Map block index within a file to block index within device.

This operation makes sense only for filesystems that use block devices, and is called only when the mount options contains blkdev.

Fields of Bmap

ino: u64block: u64blocksize: u32reply: ReplyBmap

Fallocate

Allocate requested space to an opened file.

Fields of Fallocate

ino: u64fh: u64offset: u64length: u64mode: u32reply: ReplyEmpty

CopyFileRange

Copy a range of data from an opened file to another.

Fields of CopyFileRange

ino_in: u64off_in: u64fh_in: u64ino_out: u64off_out: u64fh_out: u64len: u64flags: u64reply: ReplyWrite

Poll

Poll for readiness.

When kh is not None, the filesystem should send the notification about I/O readiness to the kernel.

Fields of Poll

ino: u64fh: u64events: u32kh: Option<u64>reply: ReplyPoll