Trait polyfuse::reply::Reply

pub trait Reply {
    fn collect_bytes<'a, T: ?Sized>(&'a self, collector: &mut T)
    where
        T: Collector<'a>;
}

A trait that represents the data structure contained in the reply sent to the kernel.

The trait is roughly a generalization of AsRef<[u8]>, representing scattered bytes that are not necessarily in contiguous memory space.

Required methods

fn collect_bytes<'a, T: ?Sized>(&'a self, collector: &mut T) where
T: Collector<'a>,

Collect the scattered bytes in the collector.

Loading content...