polyfuse::bytes

Trait Bytes

pub trait Bytes {
    // Required methods
    fn size(&self) -> usize;
    fn count(&self) -> usize;
    fn fill_bytes<'a>(&'a self, dst: &mut dyn FillBytes<'a>);
}

A trait that represents a collection of bytes.

The role of this trait is similar to Buf provided by bytes crate, but it focuses on the situation where all byte chunks are written in a single operation. This difference is due to the requirement of FUSE kernel driver that all data in a reply message must be passed in a single write(2) syscall.

Required Methods

fn size(&self) -> usize

Return the total amount of bytes contained in this data.

fn count(&self) -> usize

Return the number of byte chunks.

fn fill_bytes<'a>(&'a self, dst: &mut dyn FillBytes<'a>)

Fill with potentially multiple slices in this data.

This method corresonds to Buf::bytes_vectored, except that the number of byte chunks is acquired from Bytes::count and the implementation needs to add all chunks in dst.