Module polyfuse::op

Structs

Access
Check file access permissions.

Bmap
Map block index within a file to block index within device.

CopyFileRange
Copy a range of data from an opened file to another.

Create
Create and open a file.

DecodeError

Fallocate
Allocate requested space.

Flock
Acquire, modify or release a BSD file lock.

Flush
Close a file descriptor.

Forget
A forget information.

Forgets
A set of forget information removed from the kernel's internal caches.

Fsync
Synchronize the file contents.

Fsyncdir
Synchronize the directory contents.

Getattr
Get file attributes.

Getlk
Test for a POSIX file lock.

Getxattr
Get an extended attribute.

Interrupt
Interrupt a previous FUSE request.

Link
Create a hard link.

Listxattr
List extended attribute names.

LockOwner
The identifier for locking operations.

Lookup
Lookup a directory entry by name.

Mkdir
Create a directory node.

Mknod
Create a file node.

NotifyReply
A reply to a NOTIFY_RETRIEVE notification.

Open
Open a file.

Opendir
Open a directory.

Poll
Poll for readiness.

Read
Read data from a file.

Readdir
Read contents from an opened directory.

Readlink
Read a symbolic link.

Release
Release an opened file.

Releasedir
Release an opened directory.

Removexattr
Remove an extended attribute.

Rename
Rename a file.

Rmdir
Remove a directory.

Setattr
Set file attributes.

Setlk
Acquire, modify or release a POSIX file lock.

Setxattr
Set an extended attribute.

Statfs
Get the filesystem statistics.

Symlink
Create a symbolic link.

Unlink
Remove a file.

Write
Write data to a file.

Enums

Operation
The kind of filesystem operation requested by the kernel.

ReaddirMode

SetAttrTime
The time value requested to be set.