Struct polyfuse::KernelConfig

pub struct KernelConfig { /* fields omitted */ }

Parameters for setting up the connection with FUSE driver and the kernel side behavior.

Implementations

impl KernelConfig

pub fn async_read(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports asynchronous read requests.

Enabled by default.

pub fn atomic_o_trunc(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports the O_TRUNC open flag.

Enabled by default.

pub fn auto_inval_data(&mut self, enabled: bool) -> &mut Self

Specify that the kernel check the validity of attributes on every read.

Enabled by default.

pub fn async_dio(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports asynchronous direct I/O submission.

Enabled by default.

pub fn parallel_dirops(&mut self, enabled: bool) -> &mut Self

Specify that the kernel supports parallel directory operations.

Enabled by default.

pub fn handle_killpriv(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem is responsible for unsetting setuid and setgid bits when a file is written, truncated, or its owner is changed.

Enabled by default.

pub fn posix_locks(&mut self, enabled: bool) -> &mut Self

The filesystem supports the POSIX-style file lock.

pub fn flock_locks(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports the flock handling.

pub fn export_support(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports lookups of "." and "..".

pub fn dont_mask(&mut self, enabled: bool) -> &mut Self

Specify that the kernel should not apply the umask to the file mode on create operations.

pub fn writeback_cache(&mut self, enabled: bool) -> &mut Self

Specify that the kernel should enable writeback caching.

pub fn posix_acl(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports POSIX access control lists.

pub fn readdirplus(&mut self, enabled: bool) -> &mut Self

Specify that the filesystem supports readdirplus operations.

pub fn readdirplus_auto(&mut self, enabled: bool) -> &mut Self

Indicates that the kernel uses the adaptive readdirplus.

This option is meaningful only if readdirplus is enabled.

pub fn max_readahead(&mut self, value: u32) -> &mut Self

Set the maximum readahead.

pub fn max_write(&mut self, value: u32) -> &mut Self

Set the maximum size of the write buffer.

Panic

It causes an assertion panic if the setting value is less than the absolute minimum.

pub fn max_background(&mut self, max_background: u16) -> &mut Self

Return the maximum number of pending background requests.

pub fn congestion_threshold(&mut self, mut threshold: u16) -> &mut Self

Set the threshold number of pending background requests that the kernel marks the filesystem as congested.

If the setting value is 0, the value is automatically calculated by using max_background.

Panics

It cause a panic if the setting value is greater than max_background.

pub fn time_gran(&mut self, time_gran: u32) -> &mut Self

Set the timestamp resolution supported by the filesystem.

The setting value has the nanosecond unit and should be a power of 10.

The default value is 1.